+++ src/utils/exportFiles.ts
@@ -9,8 +9,8 @@ export function generateTsConfig() {
   return JSON.stringify({
     compilerOptions: {
       module: 'commonjs',
-      target: 'ES2019',
-      lib: ['ES2019'],
+      target: 'esnext',
+      lib: ['esnext'],
       outDir: 'dist',
       sourceMap: true,
       strict: true,
@@ -81,7 +81,7 @@ export function generateCodeWorkspaceFile(subgraphNames: string[]) {
   return JSON.stringify(codeWorkspaceFile);
 }
 
-export function generateSubgraphAction(graph: string) {
+export function generateSubgraphAction(graph: string, port: number) {
   return `
 name: Publish Schema to Apollo Studio
 
@@ -100,6 +100,7 @@ jobs:
       APOLLO_VCS_COMMIT: \${{ github.event.pull_request.head.sha }}
       APOLLO_KEY: \${{ secrets.APOLLO_KEY }}
       APOLLO_GRAPH_REF: ${graph}@production
+      ROUTING_URL: http://localhost:${port}
     steps: 
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v2
@@ -120,7 +121,6 @@ const federatedServerPackageJson: any = {
     '@apollo/federation': 'latest',
     'apollo-datasource': 'latest',
     'apollo-server': 'latest',
-    'graphql-tools': 'latest',
     faker: 'latest',
   },
   devDependencies: {
@@ -129,15 +129,10 @@ const federatedServerPackageJson: any = {
   },
 };
 
-export function generateJsFederatedServerPackageJson(
-  serviceName: string,
-  port: number = 4001,
-  graphRef: string = 'GRAPH_REF_NAME',
-) {
+export function generateJsFederatedServerPackageJson(serviceName: string) {
   federatedServerPackageJson.name = serviceName;
-  federatedServerPackageJson.scripts[
-    'publish-schema'
-  ] = `rover subgraph publish ${graphRef}@\${APOLLO_GRAPH_VARIANT} --name=${serviceName} --schema=schema.graphql --routing-url=http://localhost:${port} --convert`;
+  federatedServerPackageJson.scripts['publish-schema'] =
+    '    "publish-schema": "rover subgraph publish ${APOLLO_GRAPH_REF} --name=locations --schema=./schema.graphql --routing-url=${ROUTING_URL}';
   return JSON.stringify(federatedServerPackageJson);
 }
 
@@ -146,61 +141,55 @@ export function generateTsFederatedServerPackageJson(serviceName: string) {
   base.name = serviceName;
   base.devDependencies = {
     '@types/node': 'latest',
-    '@types/node-fetch': 'latest',
-    apollo: 'latest',
     typescript: 'latest',
-    copyfiles: '2.3.0',
   };
   base.scripts = {
-    'local-schema-validation':
-      "apollo service:check --variant=$(grep APOLLO_GRAPH_VARIANT .env | cut -d '=' -f2) --serviceName=${serviceName} --localSchemaFile=src/schema.graphql",
-    'local-schema-push-local':
-      "apollo service:push --variant=$(grep APOLLO_GRAPH_VARIANT .env | cut -d '=' -f2) --serviceName=${serviceName} --serviceURL=$(grep SERVICE_URL .env | cut -d '=' -f2) --localSchemaFile=src/schema.graphql",
-    'copy-schema': 'copyfiles -u 1 src/schema.graphql dist',
-    postinstall: 'npm run copy-schema && tsc --build tsconfig.json',
+    postinstall: 'tsc --build tsconfig.json',
     start: 'node dist/index.js',
     watch: 'tsc --build tsconfig.json --watch',
+    'publish-schema':
+      'rover subgraph publish ${APOLLO_GRAPH_REF} --name=locations --schema=./schema.graphql --routing-url=${ROUTING_URL}',
   };
 
   return JSON.stringify(base);
 }
 
-const gatewayPackageJson: any = {
-  name: 'graphql-gateway',
-  scripts: {
-    compose:
-      'rover supergraph compose --config supergraph.yaml  > supergraph-schema.graphql',
-    start: 'node src/index.js',
-    postinstall: 'npm run compose',
-  },
-  main: 'src/index.js',
-  dependencies: {
-    '@apollo/gateway': 'latest',
-    'apollo-server': 'latest',
-  },
-  devDependencies: {
-    '@apollo/rover': 'latest',
-    dotenv: 'latest',
-  },
+//TODO: Federation 2 packages to be installed for router things
+const gatewayPackageJson: any = (version: string) => {
+  return {
+    name: 'graphql-gateway',
+    scripts: {
+      compose:
+        'rover supergraph compose --config supergraph.yaml  > supergraph-schema.graphql',
+      start: 'node src/index.js',
+      postinstall: 'npm run compose',
+    },
+    main: 'src/index.js',
+    dependencies: {
+      '@apollo/gateway': version == '2' ? '2.0.0-alpha.1' : 'latest',
+      'apollo-server': 'latest',
+      dotenv: 'latest',
+    },
+    devDependencies: {
+      '@apollo/rover': 'latest',
+    },
+  };
 };
 
-export function generateJsgatewayPackageJson() {
-  return JSON.stringify(gatewayPackageJson);
+export function generateJsgatewayPackageJson(version: string = '1') {
+  return JSON.stringify(gatewayPackageJson(version));
 }
 
-export function generateTsgatewayPackageJson() {
-  const base = gatewayPackageJson;
+export function generateTsgatewayPackageJson(version: string = '1') {
+  const base = gatewayPackageJson(version);
   base.devDependencies = {
     '@types/node': 'latest',
-    '@types/node-fetch': 'latest',
-    apollo: 'latest',
     typescript: 'latest',
-    dotenv: 'latest',
   };
   base.scripts = {
-    'list-services':
-      "apollo service:list --variant=$(grep APOLLO_GRAPH_VARIANT .env | cut -d '=' -f2)",
-    postinstall: 'tsc --build tsconfig.json',
+    ...base.scripts,
+    build: 'tsc --build tsconfig.json',
+    postinstall: 'npm run build && npm run compose',
     start: 'node dist/index.js',
     watch: 'tsc --build tsconfig.json --watch',
   };
@@ -219,6 +208,21 @@ const resolvers = {
       (resolvers += `\t${entity}: {\n\t\t__resolveReference(parent, args) {\n\t\t\treturn { ...parent }\n\t\t}\n\t},\n`),
   );
   resolvers += '}\nmodule.exports = resolvers;';
+
+  return resolvers;
+}
+
+export function generateTsFederatedResolvers(schema: string) {
+  let resolvers = `// Documentation on resolving entities:
+//  https://www.apollographql.com/docs/federation/entities/#resolving
+export const resolvers = {
+`;
+  const entities = extractEntityNames(schema);
+  entities.forEach(
+    (entity) =>
+      (resolvers += `\t${entity}: {\n\t\t__resolveReference(parent, args) {\n\t\t\treturn { ...parent }\n\t\t}\n\t},\n`),
+  );
+  resolvers += '}';
 
   return resolvers;
 }
@@ -242,20 +246,13 @@ export function generateJsFederatedServerTemplate(
   return `const { resolve } = require('path');
 const { readFileSync } = require('fs');
 const { gql, ApolloServer } = require('apollo-server');
-const { buildFederatedSchema } = require('@apollo/federation');
-const { addMocksToSchema } = require('@graphql-tools/mock');
+const { buildSubgraphSchema } = require('@apollo/federation');
 
 const mocks = require('./mocks.js');
 const resolvers = require('./resolvers.js');
 const typeDefs = gql(readFileSync(resolve(__dirname, "..","schema.graphql"), { encoding: "utf8" }));
-const federatedSchema = buildFederatedSchema({ typeDefs, resolvers });
-const schema = addMocksToSchema({
-  schema: federatedSchema,
-  mocks,
-  preserveResolvers: true,
-});
-
-const server = new ApolloServer({ schema });
+const schema = buildSubgraphSchema({ typeDefs, resolvers });
+const server = new ApolloServer({ schema, mocks });
     
 const port = process.env.PORT || ${port};
 server.listen({ port }).then(({ url }) => {
@@ -270,28 +267,28 @@ export function generateTsFederatedServerTemplate(
   return `import { resolve } from 'path';
 import { readFileSync } from 'fs';
 import { gql, ApolloServer } from 'apollo-server';
-import { buildFederatedSchema } from '@apollo/federation';
+import { buildSubgraphSchema } from '@apollo/federation';
+import { resolvers } from './resolvers';
 
-const typeDefs = gql(readFileSync(resolve(__dirname, "./schema.graphql"), { encoding: "utf8" }));
-const server = new ApolloServer({
-    schema: buildFederatedSchema([{ typeDefs }]),
-    mocks: true,
-    mockEntireSchema: false,
-    engine: false,
-    });
+const mocks = require('./mocks.js');
+const typeDefs = gql(readFileSync(resolve(__dirname, "..","schema.graphql"), { encoding: "utf8" }));
+const schema = buildSubgraphSchema({ typeDefs, resolvers });
+const server = new ApolloServer({ schema, mocks });
     
 const port = process.env.PORT || ${port};
 server.listen({ port }).then(({ url }) => {
-    log(\`🚀 ${serviceName} service ready at \${url}\`);
+    console.log(\`🚀 Subgraph ${serviceName} ready at \${url}\`);
 });`;
 }
 
 export function generateJsGatewayTempalte() {
-  return `require('dotenv').config();
-const { resolve } = require('path');
+  return `const { resolve } = require('path');
 const { readFileSync } = require('fs');
 const { ApolloServer } = require('apollo-server');
 const { ApolloGateway } = require('@apollo/gateway');
+
+const isProd = process.env.NODE_ENV === "production";
+if(!isProd)require('dotenv').config();
 
 let gateway;
 if(process.env.APOLLO_KEY && process.env.APOLLO_GRAPH_REF) {
@@ -301,10 +298,7 @@ if(process.env.APOLLO_KEY && process.env.APOLLO_GRAPH_REF) {
   const supergraphSdl = readFileSync(resolve(__dirname, '..', 'supergraph-schema.graphql'), { encoding: "utf8" });
   gateway = new ApolloGateway({ supergraphSdl });
 }
-const server = new ApolloServer({
-    gateway,
-    subscriptions: false
-});
+const server = new ApolloServer({ gateway });
     
 const port = process.env.PORT ||4000;
 server.listen({ port }).then(({ url }) => {
@@ -313,21 +307,27 @@ server.listen({ port }).then(({ url }) => {
 }
 
 export function generateTsGatewayTempalte() {
-  return `import { ApolloServer } from 'apollo-server';
+  return `import { resolve } from 'path';
+import { readFileSync } from 'fs';
+import { ApolloServer } from 'apollo-server';
 import { ApolloGateway } from '@apollo/gateway';
 
 const isProd = process.env.NODE_ENV === "production";
 if(!isProd)require('dotenv').config();
 
-const gateway = new ApolloGateway({debug: !isProd});
-const server = new ApolloServer({
-    gateway,
-    subscriptions: false
-});
+let gateway;
+if(process.env.APOLLO_KEY && process.env.APOLLO_GRAPH_REF) {
+  //Default to Apollo Managed Federation
+  gateway = new ApolloGateway();
+} else {
+  const supergraphSdl = readFileSync(resolve(__dirname, '..', 'supergraph-schema.graphql'), { encoding: "utf8" });
+  gateway = new ApolloGateway({ supergraphSdl });
+}
+const server = new ApolloServer({ gateway });
     
-const port = process.env.PORT || ${StateManager.settings_gatewayServerPort};
+const port = process.env.PORT ||4000;
 server.listen({ port }).then(({ url }) => {
-    log(\`🚀 Gateway ready at \${url}\`);
+    console.log(\`🚀 Graph Router ready at \${url}\`);
 });`;
 }